refine code. move azure tts to tool, refactor actions
