+++ metagpt/schema.py
@@ -18,6 +18,7 @@ import asyncio
 import json
 import os.path
 import uuid
+from abc import ABC
 from asyncio import Queue, QueueEmpty, wait_for
 from json import JSONDecodeError
 from pathlib import Path
@@ -265,7 +266,7 @@ class MessageQueue:
 T = TypeVar("T", bound="BaseModel")
 
 
-class BaseContext(BaseModel):
+class BaseContext(BaseModel, ABC):
     @classmethod
     @handle_exception
     def loads(cls: Type[T], val: str) -> Optional[T]:
