+++ metagpt/tools/azure_tts.py
@@ -7,19 +7,16 @@
 """
 from azure.cognitiveservices.speech import AudioConfig, SpeechConfig, SpeechSynthesizer
 
-from metagpt.actions.action import Action
-from metagpt.config import Config
+from metagpt.config import CONFIG
 
 
-class AzureTTS(Action):
-    def __init__(self, name, context=None, llm=None):
-        super().__init__(name, context, llm)
-        self.config = Config()
+class AzureTTS:
+    """https://learn.microsoft.com/zh-cn/azure/cognitive-services/speech-service/language-support?tabs=tts#voice-styles-and-roles"""
 
-    # Parameters reference: https://learn.microsoft.com/zh-cn/azure/cognitive-services/speech-service/language-support?tabs=tts#voice-styles-and-roles
-    def synthesize_speech(self, lang, voice, role, text, output_file):
-        subscription_key = self.config.get("AZURE_TTS_SUBSCRIPTION_KEY")
-        region = self.config.get("AZURE_TTS_REGION")
+    @classmethod
+    def synthesize_speech(cls, lang, voice, role, text, output_file):
+        subscription_key = CONFIG.get("AZURE_TTS_SUBSCRIPTION_KEY")
+        region = CONFIG.get("AZURE_TTS_REGION")
         speech_config = SpeechConfig(subscription=subscription_key, region=region)
 
         speech_config.speech_synthesis_voice_name = voice
@@ -41,5 +38,5 @@ class AzureTTS(Action):
 
 
 if __name__ == "__main__":
-    azure_tts = AzureTTS("azure_tts")
+    azure_tts = AzureTTS()
     azure_tts.synthesize_speech("zh-CN", "zh-CN-YunxiNeural", "Boy", "Hello, I am Kaka", "output.wav")
