+++ tests/metagpt/actions/test_detail_mining.py
@@ -3,20 +3,26 @@
 """
 @Time    : 2023/9/13 00:26
 @Author  : fisherdeng
-@File    : test_detail_mining.py
+@File    : test_generate_questions.py
 """
 import pytest
 
-from metagpt.actions.detail_mining import DetailMining
+from metagpt.actions.generate_questions import GenerateQuestions
 from metagpt.logs import logger
+
+context = """
+## topic
+如何做一个生日蛋糕
+
+## record
+我认为应该先准备好材料，然后再开始做蛋糕。
+"""
 
 
 @pytest.mark.asyncio
-async def test_detail_mining():
-    topic = "如何做一个生日蛋糕"
-    record = "我认为应该先准备好材料，然后再开始做蛋糕。"
-    detail_mining = DetailMining("detail_mining")
-    rsp = await detail_mining.run(topic=topic, record=record)
+async def test_generate_questions():
+    detail_mining = GenerateQuestions()
+    rsp = await detail_mining.run(context)
     logger.info(f"{rsp.content=}")
 
     assert "Questions" in rsp.content
